requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
flask-sqlalchemy==3.0.2
SQLAlchemy==1.4.46
python-dotenv==0.21.0
psycopg2-binary==2.9.5